Extract sub-frame controls.
