+++ src/main/java/oripa/gui/presenter/main/UIPanelPresenter.java
@@ -25,26 +25,16 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import oripa.domain.cptool.TypeForChange;
-import oripa.domain.creasepattern.CreasePattern;
-import oripa.domain.fold.EstimationResultRules;
-import oripa.domain.fold.halfedge.OrigamiModel;
 import oripa.domain.paint.AngleStep;
-import oripa.domain.paint.PaintContext;
 import oripa.gui.presenter.creasepattern.TypeForChangeContext;
 import oripa.gui.presenter.main.logic.GridDivNumPresentationLogic;
-import oripa.gui.presenter.main.logic.ModelComputationFacade.ComputationResult;
 import oripa.gui.presenter.main.logic.ModelComputationFacade.ComputationType;
-import oripa.gui.presenter.main.logic.ModelComputationFacadeFactory;
-import oripa.gui.presenter.main.logic.ModelIndexChangeListenerPutter;
+import oripa.gui.presenter.main.logic.SubFramePresentationLogic;
 import oripa.gui.presenter.main.logic.UIPanelPaintMenuListenerRegistration;
+import oripa.gui.presenter.main.logic.ValuePanelPresentationLogic;
 import oripa.gui.presenter.plugin.GraphicMouseActionPlugin;
-import oripa.gui.view.FrameView;
-import oripa.gui.view.estimation.EstimationResultFrameView;
 import oripa.gui.view.main.PainterScreenSetting;
-import oripa.gui.view.main.SubFrameFactory;
 import oripa.gui.view.main.UIPanelView;
-import oripa.gui.view.model.ModelViewFrameView;
-import oripa.util.MathUtil;
 
 /**
  * @author OUCHI Koji
@@ -54,13 +44,11 @@ public class UIPanelPresenter {
 	private static final Logger logger = LoggerFactory.getLogger(UIPanelPresenter.class);
 
 	private final UIPanelView view;
-	private final SubFrameFactory subFrameFactory;
 
+	private final SubFramePresentationLogic subFramePresentationLogic;
 	private final UIPanelPaintMenuListenerRegistration paintMenuListenerRegistration;
 	private final GridDivNumPresentationLogic gridDivNumPresentationLogic;
-
-	private final ModelIndexChangeListenerPutter modelIndexChangeListenerPutter;
-	private final SubFramePresenterFactory subFramePresenterFactory;
+	private final ValuePanelPresentationLogic valuePanelPresentationLogic;
 
 	private final TypeForChange[] alterLineComboDataFrom = {
 			TypeForChange.EMPTY, TypeForChange.MOUNTAIN, TypeForChange.VALLEY, TypeForChange.UNASSIGNED,
@@ -74,35 +62,21 @@ public class UIPanelPresenter {
 
 	private final PainterScreenSetting mainScreenSetting;
 
-	private final PaintContext paintContext;
-
-	private final ModelComputationFacadeFactory computationFacadeFactory;
-	private ComputationResult computationResult;
-
-	private String lastResultFilePath;
-
 	public UIPanelPresenter(final UIPanelView view,
-			final SubFrameFactory subFrameFactory,
-			final SubFramePresenterFactory subFramePresenterFactory,
+			final SubFramePresentationLogic subFramePresentationLogic,
 			final UIPanelPaintMenuListenerRegistration paintMenuListenerRegistration,
 			final GridDivNumPresentationLogic gridDivNumPresentationLogic,
-			final ModelIndexChangeListenerPutter modelIndexChangeListenerPutter,
-			final ModelComputationFacadeFactory computationFacadeFactory,
+			final ValuePanelPresentationLogic valuePanelPresentationLogic,
 			final TypeForChangeContext typeForChangeContext,
-			final PaintContext paintContext,
 			final PainterScreenSetting mainScreenSetting) {
 
 		this.view = view;
-		this.subFrameFactory = subFrameFactory;
+
+		this.subFramePresentationLogic = subFramePresentationLogic;
 
 		this.paintMenuListenerRegistration = paintMenuListenerRegistration;
 		this.gridDivNumPresentationLogic = gridDivNumPresentationLogic;
-
-		this.modelIndexChangeListenerPutter = modelIndexChangeListenerPutter;
-		this.subFramePresenterFactory = subFramePresenterFactory;
-		this.computationFacadeFactory = computationFacadeFactory;
-
-		this.paintContext = paintContext;
+		this.valuePanelPresentationLogic = valuePanelPresentationLogic;
 
 		this.mainScreenSetting = mainScreenSetting;
 
@@ -167,128 +141,15 @@ public class UIPanelPresenter {
 		// ------------------------------------------------------------
 		// fold
 
-		view.addCheckWindowButtonListener(this::showCheckerWindow);
-		view.setModelComputationListener(this::computeModels);
-		view.setShowFoldedModelWindowsListener(this::showFoldedModelWindows);
+		view.addCheckWindowButtonListener(subFramePresentationLogic::showCheckerWindow);
+		view.setModelComputationListener(subFramePresentationLogic::computeModels);
+		view.setShowFoldedModelWindowsListener(subFramePresentationLogic::showFoldedModelWindows);
 	}
 
 	/**
 	 * Updates text fields' format setting based on eps in context.
 	 */
 	public void updateValuePanelFractionDigits() {
-		view.setValuePanelFractionDigits(
-				computeValuePanelFractionDigits(paintContext.getPointEps()),
-				computeValuePanelFractionDigits(MathUtil.angleDegreeEps()));
-	}
-
-	private int computeValuePanelFractionDigits(final double eps) {
-		// 1 digit is added for precision.
-		return (int) Math.floor(Math.abs(Math.log10(eps))) + 1;
-	}
-
-	/**
-	 * display window with foldability checks
-	 */
-	private void showCheckerWindow() {
-		var frame = subFrameFactory.createFoldabilityFrame((FrameView) view.getTopLevelView());
-		var presenter = subFramePresenterFactory.createFoldabilityCheckFrameViewPresenter(
-				frame,
-				paintContext.getCreasePattern(),
-				paintContext.getPointEps());
-
-		presenter.setViewVisible(true);
-	}
-
-	private void showCheckerWindow(final OrigamiModel origamiModel, final EstimationResultRules estimationRules) {
-		var frame = subFrameFactory.createFoldabilityFrame((FrameView) view.getTopLevelView());
-		var presenter = subFramePresenterFactory.createFoldabilityCheckFrameViewPresenter(
-				frame,
-				paintContext.getCreasePattern(),
-				origamiModel,
-				estimationRules,
-				paintContext.getPointEps());
-
-		presenter.setViewVisible(true);
-	}
-
-	private void computeModels() {
-		var modelComputation = computationFacadeFactory.createModelComputationFacade(
-				view,
-				paintContext.getPointEps());
-
-		CreasePattern creasePattern = paintContext.getCreasePattern();
-
-		var origamiModels = modelComputation.buildOrigamiModels(creasePattern);
-
-		computationResult = modelComputation.computeModels(
-				origamiModels,
-				getComputationType());
-	}
-
-	private ComputationType getComputationType() {
-		return ComputationType.fromString(view.getComputationType()).get();
-	}
-
-	private void showFoldedModelWindows() {
-		var parent = (FrameView) view.getTopLevelView();
-
-		if (!computationResult.allLocallyFlatFoldable()) {
-			view.showLocalFlatFoldabilityViolationMessage();
-			showCheckerWindow();
-			return;
-		}
-
-		var origamiModels = computationResult.origamiModels();
-		var foldedModels = computationResult.foldedModels();
-
-		ModelViewFrameView modelViewFrame = subFrameFactory.createModelViewFrame(parent,
-				view.getPaperDomainOfModelChangeListener());
-
-		var modelViewPresenter = subFramePresenterFactory.createModelViewFramePresenter(
-				modelViewFrame,
-				origamiModels,
-				paintContext.getPointEps());
-		modelViewPresenter.setViewVisible(true);
-
-		EstimationResultFrameView resultFrame = null;
-
-		if (getComputationType().isLayerOrdering()) {
-			if (!computationResult.allGloballyFlatFoldable()) {
-				// wrong crease pattern exists.
-				view.showNoAnswerMessage();
-				logger.debug("estimation rules: {}", computationResult.getEstimationResultRules());
-				showCheckerWindow(
-						computationResult.getMergedOrigamiModel(),
-						computationResult.getEstimationResultRules());
-			} else {
-				logger.info("foldable layer layout is found.");
-
-				resultFrame = subFrameFactory.createResultFrame(parent);
-
-				resultFrame.setColors(
-						view.getEstimationResultFrontColor(),
-						view.getEstimationResultBackColor());
-				resultFrame.setSaveColorsListener(view.getEstimationResultSaveColorsListener());
-				// resultFrame.repaint();
-
-				var resultFramePresenter = subFramePresenterFactory.createEstimationResultFramePresenter(
-						resultFrame,
-						foldedModels,
-						paintContext.getPointEps(),
-						lastResultFilePath,
-						path -> lastResultFilePath = path);
-
-				resultFramePresenter.setViewVisible(true);
-			}
-		}
-
-		putModelIndexChangeListener(modelViewFrame, resultFrame);
-
-		modelViewPresenter.setViewVisible(true);
-	}
-
-	private void putModelIndexChangeListener(final ModelViewFrameView modelViewFrame,
-			final EstimationResultFrameView resultFrame) {
-		modelIndexChangeListenerPutter.put(modelViewFrame, resultFrame);
+		valuePanelPresentationLogic.updateValuePanelFractionDigits();
 	}
 }
